# api-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "ids",
            "required": false,
            "in": "query",
            "description": "Filter by user IDs",
            "schema": {
              "example": [
                "550e8400-e29b-41d4-a716-446655440000"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "emails",
            "required": false,
            "in": "query",
            "description": "Filter by email addresses",
            "schema": {
              "example": [
                "john@example.com",
                "jane@example.com"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "firstNames",
            "required": false,
            "in": "query",
            "description": "Filter by first names",
            "schema": {
              "example": [
                "John",
                "Jane"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "lastNames",
            "required": false,
            "in": "query",
            "description": "Filter by last names",
            "schema": {
              "example": [
                "Doe",
                "Smith"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "emailContains",
            "required": false,
            "in": "query",
            "description": "Search for emails containing this string",
            "schema": {
              "example": "gmail.com",
              "type": "string"
            }
          },
          {
            "name": "firstNameContains",
            "required": false,
            "in": "query",
            "description": "Search for first names containing this string",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "name": "lastNameContains",
            "required": false,
            "in": "query",
            "description": "Search for last names containing this string",
            "schema": {
              "example": "Doe",
              "type": "string"
            }
          },
          {
            "name": "phoneContains",
            "required": false,
            "in": "query",
            "description": "Search for phones containing this string",
            "schema": {
              "example": "+380",
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": true,
            "in": "query",
            "description": "Sort order (field:direction)",
            "schema": {
              "default": "createdAt:desc",
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "roles",
            "required": false,
            "in": "query",
            "description": "Filter by user roles",
            "schema": {
              "example": [
                "User",
                "Admin"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "User",
                  "Admin",
                  "Manager"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of users",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserPublicEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User was successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot create user. Invalid data was provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error has occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UserController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The uuid of the user to be found.",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          },
          {
            "name": "emailContains",
            "required": false,
            "in": "query",
            "description": "Search for email containing this string",
            "schema": {
              "example": "gmail.com",
              "type": "string"
            }
          },
          {
            "name": "firstNameContains",
            "required": false,
            "in": "query",
            "description": "Search for first name containing this string",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "name": "lastNameContains",
            "required": false,
            "in": "query",
            "description": "Search for last name containing this string",
            "schema": {
              "example": "Doe",
              "type": "string"
            }
          },
          {
            "name": "phoneContains",
            "required": false,
            "in": "query",
            "description": "Search for phones containing this string",
            "schema": {
              "example": "+380",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "example": "Admin",
              "type": "string",
              "enum": [
                "User",
                "Admin",
                "Manager"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user with requested id.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "The user with the requested id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The uuid of the user to be updated",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User was successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "The user with the requested id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot update user. Invalid data was provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The id of the user to be deleted",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User was successfully removed.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "The user with the requested id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/orders": {
      "get": {
        "operationId": "UserController_findAllUserOrders",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The uuid of the user to be found.",
            "schema": {
              "example": "23fbed56-1bb9-40a0-8977-2dd0f0c6c31f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of orders",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "404": {
            "description": "The user with the requested id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-Commerce Order Management API",
    "description": "RESTful API for managing users and orders in an e-commerce system",
    "version": "1.0.0",
    "contact": {
      "name": "Contact API Support",
      "url": "https://t.me/kalyonekenobe",
      "email": "oleksandr.ihumnov@ukma.edu.ua"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.ukmashop.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponseEntity": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "Invalid email"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "UserPublicEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user (UUID v4)",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "description": "User email address used for login and communication",
            "format": "email",
            "example": "user@ukmashop.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phone": {
            "type": "object",
            "description": "User phone number in international format (10–15 digits, optional + prefix)",
            "example": "+1234567890"
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user that determines access permissions",
            "enum": [
              "User",
              "Admin",
              "Manager"
            ],
            "example": "User"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the user account is currently active",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the user account was created (ISO 8601 format)",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the user account was last updated (ISO 8601 format)",
            "example": "2024-01-20T14:45:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phone",
          "role",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApiErrorResponseEntity": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-02-15T21:24:28.191Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/endpoint"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occured on server"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null,
            "additionalProperties": {}
          },
          "errors": {
            "type": "object",
            "example": {
              "server": [
                "An error occured on server"
              ]
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "timestamp",
          "path",
          "success",
          "message",
          "data",
          "errors"
        ]
      },
...
      "PaginatedApiResponseEntity": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "Invalid email"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalCount": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "success",
          "data",
          "page",
          "limit",
          "totalCount"
        ]
      },
...
    }
  }
}